changed theme for more light weight

+++ srcs/requirements/wordpress/tools/setup.sh
@@ -19,38 +19,36 @@ while [ $attempt -lt $MAX_RETRIES ]; do
 done
 
 
-
-wp core download --allow-root
-
-wp config create --allow-root \
-	--dbhost=$MARIADB_HOST \
-	--dbname=$MARIADB_DATABASE \
-	--dbuser=$MARIADB_USER \
-	--dbpass=$MARIADB_PASSWORD \
-	--path=/var/www/html \
-	--force
-
-wp core install --allow-root \
-	--url=https://$DOMAIN_NAME \
-	--title=$WP_TITLE \
-	--admin_user=$WP_ADMIN_USER \
-	--admin_password=$WP_ADMIN_PASSWORD \
-	--admin_email=$WP_ADMIN_EMAIL \
-	--path=/var/www/html/
-
-wp user create \
-	$WP_USER \
-	$WP_EMAIL \
-	--role=author \
-	--user_pass=$WP_PASSWORD \
-	--allow-root
-
-wp theme install inspiro --allow-root --activate
-wp plugin update --all --allow-root
-wp option update siteurl --allow-root "https://$DOMAIN_NAME"
-wp option update home --allow-root "https://$DOMAIN_NAME"
-echo "Finished installation and setup!"
-
+if [ -f "wp-config.php" ]; then
+    echo "WordPress is already installed."
+else
+	wp core download --allow-root
+	wp config create --allow-root \
+		--dbhost=$MARIADB_HOST \
+		--dbname=$MARIADB_DATABASE \
+		--dbuser=$MARIADB_USER \
+		--dbpass=$MARIADB_PASSWORD \
+		--path=/var/www/html \
+		--force
+	wp core install --allow-root \
+		--url=https://$DOMAIN_NAME \
+		--title=$WP_TITLE \
+		--admin_user=$WP_ADMIN_USER \
+		--admin_password=$WP_ADMIN_PASSWORD \
+		--admin_email=$WP_ADMIN_EMAIL \
+		--path=/var/www/html/
+	wp user create \
+		$WP_USER \
+		$WP_EMAIL \
+		--role=author \
+		--user_pass=$WP_PASSWORD \
+		--allow-root
+	wp theme install neve --allow-root --activate
+	wp plugin update --all --allow-root
+	wp option update siteurl --allow-root "https://$DOMAIN_NAME"
+	wp option update home --allow-root "https://$DOMAIN_NAME"
+	echo "Finished installation and setup!"
+fi
 
 echo "PHP engine runnning"
 exec /usr/sbin/php-fpm7.4 -F
